fix(rule): retain module-scoped state resets
